// apps/branch-services/src/app/(main)/new-requests/utils/enums.ts
export const enum TransactionType {
  ALL = 'ALL',
  PAYA = 'PAYA',
  SATNA = 'SATNA',
  LOCAL = 'LOCAL',
  AUTO = 'AUTO',
  INTRA_BANK = 'INTRA_BANK',
}

export enum NEW_REQUESTS_QUERY_KEYS {
  HISTORY = 'HISTORY',
  VALIDATE_REQUEST = 'VALIDATE_REQUEST',
  CHECK_VALIDATION = 'CHECK_VALIDATION',
  APPROVE_REQUEST = 'APPROVE_REQUEST',
  APPROVE_OFFLINE_ACH_REQUEST = 'APPROVE_OFFLINE_ACH_REQUEST',
  FILE_DETAILS = 'FILE_DETAILS',
  PAYMENT_ID = 'PAYMENT_ID',
  STATUSES = 'STATUSES',
  DOWNLOAD_OFFLINE_FILE_REQUEST = 'DOWNLOAD_OFFLINE_FILE_REQUEST',
  USER_FILE_DETAILS = 'USER_FILE_DETAILS',
  TRANSACTION_DETAILS = 'TRANSACTION_DETAILS',
  PAYMENT_LIMIT_INQUIRY = 'PAYMENT_LIMIT_INQUIRY',
}

export const enum StepRoute {
  VALIDATION_STEP = 'VALIDATION_STEP',
  FILE_DETAILS = 'FILE_DETAILS',
  FINAL_CONFIRMATION = 'FINAL_CONFIRMATION',
}

export enum PageKind {
  HISTORY = 'HISTORY',
  STEPPER = 'STEPPER',
}

export const enum RequestStatus {
  DEFAULT = 'DEFAULT',
  TRY_AGAIN = 'TRY_AGAIN',
  SUCCESS = 'SUCCESS',
}

export const enum Statuses {
  CONFIRMED = 'CONFIRMED',
  CREATED = 'CREATED',
  DENIED = 'DENIED',
  PARTIALLY_SUCCESS = 'PARTIALLY_SUCCESS',
  INITIATED = 'INITIATED',
  IN_PROGRESS = 'IN_PROGRESS',
  DETAILS_INSERTED = 'DETAILS_INSERTED',
  FILE_ERROR = 'FILE_ERROR',
  SUBMITTED = 'SUBMITTED',
  CANCELED = 'CANCELED',
  PROCESSING = 'PROCESSING',
  PENDING = 'PENDING',
  PAID = 'PAID',
  WRITTEN_TO_MQ = 'WRITTEN_TO_MQ',
  SUCCEED = 'SUCCEED',
  FAIL = 'FAIL',
  CONFLICT_RECORDS = 'CONFLICT_RECORDS',
  IN_OPERATION = 'IN_OPERATION',
  REVERTED = 'REVERTED',
  REVERT_FAILED = 'REVERT_FAILED',
  NONE = 'NONE',
  INACTIVE = 'INACTIVE',
  ACTIVE = 'ACTIVE',
  SUCCESS = 'SUCCESS',
  FAILURE = 'FAILURE',
  VALIDATION_SUCCESS = 'VALIDATION_SUCCESS',
  VALIDATION_FAILURE = 'VALIDATION_FAILURE',
}
export const enum RequestTypes {
  ONLINE = 'ONLINE',
  OFFLINE_ACH = 'OFFLINE_ACH',
  OFFLINE_SALARY = 'OFFLINE_SALARY',
}